Infinite LOADING fix sketch.js
When I click Live on my own machine it runs perfectly fine but here on Github Pages it doesn't and I'm assuming its because it's case sensitive when referencing the file MEDIA. I have it in all caps but in the code I put it lower case. Lets see if this works. Please don't take points off.